requirements.txt
yfinance
pandas
requests
beautifulsoup4
python-dotenv
schedule
reportlab
openpyxl
twilio

google-generativeai
markdown
